requirements.txt
pandas

jupyterlab

jedi<=0.17.2


astropy

ics
joblib
matplotlib

PyYAML

####
websockets
websocket-client

opencv-python
cv2-utils
Pillow-PIL
pyserial